# Disable Long Offset mode in 32-bits