faster decoding in 32-bits mode for long offsets

On my laptop:
Before:
```
./zstd32 -b --zstd=wlog=27 silesia.tar enwik8 -S
 3#silesia.tar       : 211984896 ->  66683478 (3.179),  97.6 MB/s , 400.7 MB/s
 3#enwik8            : 100000000 ->  35643153 (2.806),  76.5 MB/s , 303.2 MB/s
```

After:
```
./zstd32 -b --zstd=wlog=27 silesia.tar enwik8 -S
 3#silesia.tar       : 211984896 ->  66683478 (3.179),  97.4 MB/s , 435.0 MB/s
 3#enwik8            : 100000000 ->  35643153 (2.806),  76.2 MB/s , 338.1 MB/s
```

Mileage vary, depending on file, and cpu type.
A generic rule of thumb : x86 benefits less from "long-offset mode" than x64, maybe due to register pressure.
On "entropy", long-mode is _never_ a win for x86.
On my laptop, it may, depending on file and compression level
(`enwik8` typically benefits more from "long-mode" than `silesia`).

As a quick stop-gap, this patch simply disables long-offset mode in 32-bits.

Long-offset mode is currently triggered depending on `windowLog` value.
This input is probably not enough,
I suspect it would be better to look at distribution of offset codes in order to decide triggering long-offset mode or not. However, it requires to re-order decoding pipeline. That's a refinement for later.